Working on album search results + implemented remove from queue functionality
